Shimon commit. Updating the student screen so that the associated courses will be displayed on the screen, and if the student is already logged on, he will still be shown his courses.
