+++ app/src/main/AndroidManifest.xml
@@ -19,9 +19,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".StudentActivity" />
+        <activity android:name=".StudentCourseActivity" />
         <activity android:name=".TeacherActivity" />
-        <activity android:name=".StudentHistory"></activity>
+        <activity android:name=".StudentActivity"></activity>
     </application>
 
 </manifest>
