+++ app/src/main/java/com/example/android/mepo/LoginActivity.java
@@ -3,6 +3,7 @@ package com.example.android.mepo;
 
 
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
@@ -34,6 +35,7 @@ public class LoginActivity extends AppCompatActivity implements View.OnClickList
     private ProgressBar mProgressBar;
 
 
+    public static String IsStudent;
 
 
 
@@ -49,17 +51,20 @@ public class LoginActivity extends AppCompatActivity implements View.OnClickList
         //if user already logged in then pass him to his activity
         if(SharedPrefManager.getInstance(this).isLoggedIn()){
 
+            IsStudent = SharedPrefManager.getInstance(this).getUserIsStudent();
 
-            if(SharedPrefManager.getInstance(this).getUserDepartment() == null) {
+            //if(SharedPrefManager.getInstance(this).getUserDepartment() == null) {
+            if(SharedPrefManager.getInstance(this).getUserIsStudent() == null) {
                 finish();
-
+                //teacherLogin();
                 teacherCoursesList();
                 //startActivity(new Intent(this, TeacherActivity.class));              
                 return;
             }
             else {
                 finish();
-                startActivity(new Intent(this, StudentActivity.class));
+                studentCoursesList();
+                //startActivity(new Intent(this, StudentActivity.class));
                 return;
             }
 
@@ -115,10 +120,37 @@ public class LoginActivity extends AppCompatActivity implements View.OnClickList
                                                         jsonObject.getString("s_email"),
                                                         jsonObject.getString("s_department")
                                                 );
+                                        IsStudent = SharedPrefManager.getInstance(getApplicationContext()).getUserIsStudent();
 
-                                        //Get us to TeacherActivity screen
-                                        startActivity(new Intent(getApplicationContext(), StudentActivity.class));
+                                        JSONArray courses_arr;
+                                        courses_arr = jsonObject.getJSONArray("c_names");
+
+                                        ArrayList<String> list_of_courses = new ArrayList<String>();
+                                        if (courses_arr != null) {
+                                            System.out.println("courses_arr not null!!!!!!!!!!!!");
+                                            for (int i = 0; i < courses_arr.length(); i++) {
+                                                list_of_courses.add(courses_arr.getString(i));
+                                                //System.out.println(list_of_courses.get(i));
+                                            }
+                                        }else{
+                                            System.out.println("courses_arr = null!!!!!!!!!!!!");
+                                        }
+
+                                        //Get us to StudentActivity screen
+                                        Intent intent = new Intent(getBaseContext(), StudentActivity.class);
+                                        intent.putExtra("EXTRA_STUDENT_COURSES_SIZE", list_of_courses.size());
+                                        intent.putExtra("EXTRA_STUDENT_COURSES_NAME", list_of_courses);
+                                        for (int i = 0; i < list_of_courses.size(); i++) {
+                                            intent.putExtra("EXTRA_STUDENT_COURSES_NAME:"+i, list_of_courses.get(i));
+                                        }
+
+                                        startActivity(intent);
+                                        //startActivity(new Intent(getApplicationContext(), TeacherActivity.class));
                                         finish();
+
+
+                                        //startActivity(new Intent(getApplicationContext(), StudentCourseActivity.class));
+
                                     }else{
                                         Toast.makeText(
                                                 getApplicationContext(),
@@ -163,13 +195,111 @@ public class LoginActivity extends AppCompatActivity implements View.OnClickList
 
 
 
+    public void studentCoursesList(){
+
+        mProgressBar.setVisibility(View.VISIBLE);
+
+        StringRequest stringRequest = new StringRequest
+                (Request.Method.POST, Constants.URL_S_LOGIN,
+                        new Response.Listener<String>() {
+                            @Override
+                            public void onResponse(String response) {
+
+                                mProgressBar.setVisibility(View.INVISIBLE);
+                                try{
+                                    JSONObject jsonObject = new JSONObject(response);
+                                    //If there is no error message in the JSON string
+                                    if(!jsonObject.getBoolean("error")){
+
+                                        JSONArray courses_arr;
+                                        courses_arr = jsonObject.getJSONArray("c_names");
+                                        //NUM_LIST_ITEMS = courses_arr.length();
+
+                                        ArrayList<String> list_of_courses = new ArrayList<String>();
+                                        if (courses_arr != null) {
+                                            System.out.println("courses_arr not null!!!!!!!!!!!!");
+                                            System.out.println(courses_arr.length());
+                                            for (int i = 0; i < courses_arr.length(); i++) {
+                                                list_of_courses.add(courses_arr.getString(i));
+                                                //System.out.println(list_of_courses.get(i));
+                                            }
+                                        }else{
+                                            System.out.println("courses_arr = null!!!!!!!!!!!!");
+                                        }
+
+                                        System.out.println(list_of_courses);
+
+                                        //Get us to TeacherActivity screen
+                                        Intent intent = new Intent(getBaseContext(), StudentActivity.class);
+                                        intent.putExtra("EXTRA_STUDENT_COURSES_SIZE", list_of_courses.size());
+                                        intent.putExtra("EXTRA_STUDENT_COURSES_NAME", list_of_courses);
+                                        for (int i = 0; i < list_of_courses.size(); i++) {
+                                            intent.putExtra("EXTRA_STUDENT_COURSES_NAME:"+i, list_of_courses.get(i));
+                                        }
+
+                                        startActivity(intent);
+                                        //startActivity(new Intent(getApplicationContext(), TeacherActivity.class));
+                                        //finish();
+
+
+
+                                    }else{
+                                        Toast.makeText(
+                                                getApplicationContext(),
+                                                jsonObject.getString("message"),
+                                                Toast.LENGTH_LONG
+                                        ).show();
+                                    }
+
+                                } catch (JSONException e){
+                                    e.printStackTrace();
+                                }
+                            }
+                        },
+                        new Response.ErrorListener() {
+                            @Override
+                            public void onErrorResponse(VolleyError error) {
+                                mProgressBar.setVisibility(View.INVISIBLE);
+                                Toast.makeText(
+                                        getApplicationContext(),
+                                        error.getMessage(),
+                                        Toast.LENGTH_LONG
+                                ).show();
+                            }
+                        }){
+
+            //Push parameters to Request.Method.POST
+            @Override
+            protected Map<String, String> getParams() throws AuthFailureError {
+                Map<String, String> params = new HashMap<>();
+                params.put("s_id",SharedPrefManager.getInstance(getApplicationContext()).getUserId());
+                //params.put("t_password",SharedPrefManager.getInstance(getApplicationContext()).getUserPassword());
+                return params;
+            }
+        };
+
+        //Making a connection by singleton class to the database with stringRequest
+        RequestHandler.getInstance(this).addToRequestQueue(stringRequest);
+
+
+    }
+
+
+
+
+
+
+
+
 
 
 
 
     private void teacherLogin(){
+
         final String user_id = mEt_userId.getText().toString().trim();
         final String user_password = mEt_userPassword.getText().toString().trim();
+        IsStudent = null;
 
 
         mProgressBar.setVisibility(View.VISIBLE);
@@ -195,6 +325,7 @@ public class LoginActivity extends AppCompatActivity implements View.OnClickList
                                                         jsonObject.getString("t_lastName"),
                                                         jsonObject.getString("t_email")
                                                 );
+
                                         JSONArray courses_arr;
                                         courses_arr = jsonObject.getJSONArray("c_names");
 
@@ -270,6 +401,7 @@ public class LoginActivity extends AppCompatActivity implements View.OnClickList
     public void teacherCoursesList(){
 
         mProgressBar.setVisibility(View.VISIBLE);
+        IsStudent = null;
 
         StringRequest stringRequest = new StringRequest
                 (Request.Method.POST, Constants.URL_T_LOGIN,
