+++ app/src/main/java/com/example/android/mepo/StudentCourseActivity.java
@@ -1,0 +1,64 @@
+package com.example.android.mepo;
+
+import android.content.Intent;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuItem;
+import android.widget.Button;
+import android.widget.TextView;
+
+public class StudentCourseActivity extends AppCompatActivity {
+
+    private TextView  mtv_fName;
+    private Button historyButt;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.course_activity_student);
+
+        //check if user logged in or not.
+        if(!SharedPrefManager.getInstance(this).isLoggedIn()){
+            finish();
+            startActivity(new Intent(this, LoginActivity.class));
+        }
+
+        mtv_fName = findViewById(R.id.nameTextView);
+        mtv_fName.setText(SharedPrefManager.getInstance(this).getUserFName());
+        historyButt = findViewById(R.id.historyButt);
+        /*historyButt.setOnClickListener(new View.OnClickListener()
+        {
+            @Override
+            public void onClick(View v)
+            {
+                    //startActivity(new Intent(getApplicationContext(), StudentHistory.class));
+            }
+        });*/
+    }
+
+    //Responsible for the logout button
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.menu, menu);
+        return true;
+    }
+
+    //Responsible for the menu buttons, what each button does
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()){
+            case R.id.menuLogout:
+                SharedPrefManager.getInstance(this).logout();
+                finish();
+                startActivity(new Intent(this, LoginActivity.class));
+                break;
+        }
+        return true;
+    }
+
+
+
+
+
+}
