+++ app/src/main/res/layout/activity_teacher.xml
@@ -5,7 +5,7 @@
     android:padding="16dp">
 
     <android.support.v7.widget.RecyclerView
-        android:id="@+id/rv_studentCourses"
+        android:id="@+id/rv_teacherCourses"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
          />
